pyproject.toml
[tool.poetry]
name = "exoskelton"
version = "0.1.0"
description = "A bot framework for slack"
authors = ["Toshiaki Hatano <haeena@haeena.net>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/haeena/exoskelton"

include = [
    "README.md",
    "pyproject.toml",
    "LICENSE"
]

[tool.poetry.dependencies]
python = "^3.8"
slackeventsapi = "^2.1"
slackclient = "^2.3"

[tool.poetry.dev-dependencies]
ptvsd = "^4.3"
flake8 = "^3.7"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest = "^5.2"
isort = "^4.3"
coverage = "^4.5"
pytest-cov = "^2.8"
mypy = "^0.740.0"
pre-commit = "^1.20"
tox = "^3.14"
bandit = "^1.6.2"
safety = "^1.8.5"
python-taint = "^0.42"
flake8-bugbear = "^19.8.0"

[tool.black]
line-length = 80
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
known_first_party = 'exoskelton'
known_third_party = ["flask", "mypy_extensions", "ptvsd", "slack", "slackeventsapi"]
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = true
envlist = py38

[testenv]
whitelist_externals = poetry
commands =
    poetry install --no-dev -vvv
    poetry run pytest --cov=exoskelton tests/
"""

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
